Change package structure
state -> state.concrete
state.frame -> state
