+++ src/main/java/com/group/artifact/state/concrete/BookInfoState.java
@@ -1,16 +1,16 @@
-package com.group.artifact.state;
+package com.group.artifact.state.concrete;
 
 import com.group.artifact.domain.Book;
+import com.group.artifact.service.SlackService;
 import com.group.artifact.state.container.StateContainer;
-import com.group.artifact.state.frame.State;
+import com.group.artifact.state.NeedBookNameState;
 import com.group.artifact.state.proxy.ProxyWithManyBookNames;
 import com.group.artifact.vo.MessageVo;
-import com.group.artifact.service.SlackService;
 
 import java.util.List;
 
-public class ReviewUpdate extends State {
-    public ReviewUpdate(SlackService service, StateContainer stateContainer) {
+public class BookInfoState extends NeedBookNameState {
+    public BookInfoState(SlackService service, StateContainer stateContainer) {
         super(service, stateContainer);
     }
 
@@ -22,14 +22,18 @@ public class ReviewUpdate extends State {
             stateContainer.remove(messageVo);
             return "NO BOOK NAME";
         } else if (books.size() == 1) {
-            service.askReviewContents(messageVo.getChannel());
-            stateContainer.put(messageVo, new ReviewUpdateExpectContents(service, stateContainer, books.get(0).getTitle()));
-            return "ASK CONTENTS";
+            return serviceWithBookName(messageVo, books.get(0).getTitle());
         } else {
             service.selectBook(messageVo.getChannel(), books);
-            stateContainer.put(messageVo, new ProxyWithManyBookNames(books, new ReviewUpdateExpectContents(service, stateContainer)));
+            stateContainer.put(messageVo, new ProxyWithManyBookNames(books, this));
             return "MANY BOOK";
         }
     }
 
+    @Override
+    public String serviceWithBookName(MessageVo messageVo, String bookName) {
+        service.sendBookInfo(bookName, messageVo.getChannel());
+        stateContainer.remove(messageVo);
+        return "BOOK INFO";
+    }
 }
