+++ src/main/java/com/group/artifact/state/concrete/DoNothing.java
@@ -1,18 +1,18 @@
-package com.group.artifact.state;
+package com.group.artifact.state.concrete;
 
 import com.group.artifact.state.container.StateContainer;
-import com.group.artifact.state.frame.State;
+import com.group.artifact.state.State;
 import com.group.artifact.vo.MessageVo;
 import com.group.artifact.service.SlackService;
 
-public class Usage extends State {
-    public Usage(SlackService service, StateContainer stateContainer) {
+public class DoNothing extends State {
+
+    public DoNothing(SlackService service, StateContainer stateContainer) {
         super(service, stateContainer);
     }
 
     @Override
     public String doService(MessageVo messageVo) {
-        service.usage(messageVo.getChannel());
-        return "PRINT USAGE";
+        return "DO NOTHING";
     }
 }
