+++ src/main/java/com/group/artifact/state/concrete/ExitCommand.java
@@ -1,8 +1,8 @@
-package com.group.artifact.state;
+package com.group.artifact.state.concrete;
 
 import com.group.artifact.service.SlackService;
 import com.group.artifact.state.container.StateContainer;
-import com.group.artifact.state.frame.State;
+import com.group.artifact.state.State;
 import com.group.artifact.vo.MessageVo;
 
 public class ExitCommand extends State {
