+++ src/main/java/com/group/artifact/state/container/StateContainer.java
@@ -1,7 +1,7 @@
 package com.group.artifact.state.container;
 
 import com.group.artifact.service.SlackService;
-import com.group.artifact.state.frame.State;
+import com.group.artifact.state.State;
 import com.group.artifact.vo.MessageVo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
